+++ apps/native-component-list/src/screens/Calendar@Next/EventsNextScreen.tsx
@@ -4,10 +4,10 @@ import { ExpoCalendar, ExpoCalendarEvent } from 'expo-calendar/next';
 import React, { useState, useEffect } from 'react';
 import { Alert, Platform, ScrollView, StyleSheet, Text, View } from 'react-native';
 
-import Button from '../components/Button';
-import HeadingText from '../components/HeadingText';
-import ListButton from '../components/ListButton';
-import MonoText from '../components/MonoText';
+import Button from '../../components/Button';
+import HeadingText from '../../components/HeadingText';
+import ListButton from '../../components/ListButton';
+import MonoText from '../../components/MonoText';
 
 type EventRowProps = {
   event: ExpoCalendarEvent;