+++ packages/expo-calendar/android/src/main/java/expo/modules/calendar/next/records/CalendarPermissionOptions.kt
@@ -1,0 +1,10 @@
+package expo.modules.calendar.next.records
+
+import expo.modules.kotlin.records.Field
+import expo.modules.kotlin.records.Record
+import expo.modules.kotlin.types.OptimizedRecord
+
+@OptimizedRecord
+data class CalendarPermissionOptions(
+  @Field val writeOnly: Boolean = false
+) : Record