+++ packages/expo-calendar/ios/Next/ExpoCalendarPermissions.swift
@@ -34,23 +34,35 @@ public class ExpoCalendarPermissions {
     try self.checkPermissions(entity: .event)
   }
 
+  public func checkCalendarWritePermissions() throws {
+    try self.checkPermissions(
+      entity: .event,
+      requester: CalendarWriteOnlyPermissionsRequester.self
+    )
+  }
+
   public func checkRemindersPermissions() throws {
     try self.checkPermissions(entity: .reminder)
   }
 
-  private func checkPermissions(entity: EKEntityType) throws {
+  private func checkPermissions(
+    entity: EKEntityType,
+    requester requestedRequester: EXPermissionsRequester.Type? = nil
+  ) throws {
     guard let permissionsManager = appContext?.permissions else {
       throw PermissionsManagerNotFoundException()
     }
 
-    var requester: EXPermissionsRequester.Type?
-    switch entity {
-    case .event:
-      requester = CalendarPermissionsRequester.self
-    case .reminder:
-      requester = RemindersPermissionRequester.self
-    @unknown default:
-      requester = nil
+    var requester = requestedRequester
+    if requester == nil {
+      switch entity {
+      case .event:
+        requester = CalendarPermissionsRequester.self
+      case .reminder:
+        requester = RemindersPermissionRequester.self
+      @unknown default:
+        requester = nil
+      }
     }
     if let requester, !permissionsManager.hasGrantedPermission(usingRequesterClass: requester) {
       let message = requester.permissionType().uppercased()